moved generate_acess_response to helpers/responces
